Changed all ids from longs to ints.

+++ src/Action.java
@@ -4,5 +4,5 @@ public interface Action
 	public boolean enabled();
 	public void act();
 	public String getAllInfo();
-	public long id();
+	public int id();
 }
